#4 What is MVT in Django. Django Documentation
what is mvt architecture in Django ?
The MVT (Model View Template) is a software design pattern. It is a collection of three important components Model View and Template. 
The Model helps to handle database. It is a data access layer which handles the data. There is no separate controller and complete 
application is based on Model View and Template.
MVC has controller that drives both Model and View. MVT has Views for recieving HTTP request and returning HTTP response. 2. 
View tells how the user data will be presented. Templates are used in MVT for that purpose

Django Officila website Link: https://www.djangoproject.com