#7  to add two Addition numbers | GET POST Http
In this Tutorial I'm going to Teach you How to add two Additon numbers and What is GET POST. What is HTTP request and response in Django ?
What is csrf and How it's work and how to do.
Watch video from My YouTube cannel:  https://www.youtube.com/channel/UCV6vGLwmJneo7leWpgjVBDA     

What is HTTP request and response in Django ?

HTTP works as a request-response protocol between a client and server. Example: A client (browser) sends an HTTP request to the server; then 
the server returns a response to the client. The response contains status information about the request and may also contain the requested 
content.

What is the GET Method django ?
The GET method refers to a HyperText Transfer Protocol (HTTP) method that is applied while requesting information from a particular source. 
It is also used to get a specific variable derived from a group. The HTTP POST asks for input of information from the supplying browser into 
the server’s message system. Structures of HyperText Markup Language (HTML) come in either POST method or GET method. The method applied, be 
it the POST method or the GET method, settles on how form data shall be presented before the server.
To understand the attributes of the GET method or otherwise termed as HTTP GET request, here is an outlined comparison between the GET method 
and the POST methods.

what is POST request ?
In computing, POST is a request method supported by HTTP used by the World Wide Web. By design, the POST request method requests that a web 
server accepts the data enclosed in the body of the request message, most likely for storing it. It is often used when uploading a file or 
when submitting a completed web form


Differences between the POST method and the GET method ?
1. For the GET method, the parameters maintain within the server since it forms a portion of the URL (Uniform Resource Locator.). 
On the other hand, the status of parameters of the POST method of HTTP does not remain since it is not within the browser history.

2. The GET method of HTTP is capable of being bookmarked since it is within the URL but the POST method cannot be bookmarked.

3. By way of the GET method of HTTP, one can send parameter information but with a limitation to what is acceptable to the URL. 
On the other hand, one can send information to the server by POST method without any kind of hindrances, including information in 
the form of uploaded files. Information inscribed on the URL is limited to an approximate of 2000 characters under the GET method. 
The POST method does not feature such restrictions.

4. The GET method is easily meddled with, thus is not safe since the information received is also saved in the URL, meaning anyone else can 
have access to it. This attribute renders the GET method applicability at a disadvantage. On the other hand, the POST method is safer and not 
easily cracked up. POST parameters are not maintained in the server or webpages thus it is not easily accessed. It is therefore discouraged 
for one to inscribe delicate information through the GET method such as credit card pins and passcodes. Such information is detectable as 
information posted on the URL is accessible to anyone.
One can send a GET request with added php object. Here is an example of a GET request with an additional php object about a 54 year old 
George in Canada. The parameters of the example will send a GET request as follows: test.php?name=George&age=54&country=Canada